requirements.txt
torch==2.8.0
torchvision==0.23.0
psycopg2-binary==2.9.10
pandas==2.3.2
pillow==11.1.0
